Combine all hooks in a place
